+++ antcrew/__init__.py
@@ -17,6 +17,7 @@ from antcrew.models.cache import LLMCache, FileLLMCache
 from antcrew.sandbox import DockerRunner, LocalRunner, SandboxRunner
 from antcrew.sandbox import RunResult as SandboxRunResult
 from antcrew.core.run_result import RunResult
+from antcrew.checkpointers import SqliteSaver
 from antcrew.flow import load_flow, validate_flow, format_flow
 from antcrew.project import Project
 from antcrew.config import load_context, TeamContext, build_llm, build_runner
@@ -71,6 +72,7 @@ __all__ = [
     "DockerRunner",
     "RunResult",
     "SandboxRunResult",
+    "SqliteSaver",
     # Flow
     "load_flow",
     "validate_flow",